# public/api/v1/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quotd API",
    "description": "Quotd REST API for managing AI voice interviews and case study generation.",
    "version": "1.0.0",
    "contact": {
      "name": "Quotd",
      "url": "https://quotd.io"
    }
  },
  "servers": [
    {
      "url": "https://app.quotd.io/api/v1",
      "description": "Production"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key (qtd_...) as Bearer token"
      }
    },
    "schemas": {
      "Interview": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string", "format": "uuid" },
          "customer_company": { "type": "string" },
          "product_name": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "in_progress", "review_pending", "review_in_progress", "review_complete"] },
          "share_token": { "type": "string" },
          "customer_email": { "type": "string", "nullable": true },
          "interview_tone": { "type": "string", "enum": ["formal", "conversational", "technical"] },
          "interview_focus": { "type": "string", "enum": ["balanced", "roi", "technical", "storytelling"] },
          "target_audience": { "type": "string", "enum": ["general", "c_suite", "technical_buyer", "end_user", "board"] },
          "question_limit": { "type": "integer" },
          "extraction_state": { "$ref": "#/components/schemas/ExtractionState" },
          "draft_content": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "ExtractionState": {
        "type": "object",
        "properties": {
          "metrics": { "type": "array", "items": { "$ref": "#/components/schemas/Metric" } },
          "quotes": { "type": "array", "items": { "$ref": "#/components/schemas/Quote" } },
          "facts": { "type": "object", "properties": { "challenge": { "type": "string" }, "solution": { "type": "string" }, "impact": { "type": "string" } } },
          "question_count": { "type": "integer" }
        }
      },
      "Metric": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "baseline": { "type": "string", "nullable": true },
          "after": { "type": "string", "nullable": true },
          "delta": { "type": "string", "nullable": true },
          "unit": { "type": "string" },
          "timeframe": { "type": "string", "nullable": true },
          "confidence": { "type": "string", "enum": ["high", "medium", "low"] }
        }
      },
      "Quote": {
        "type": "object",
        "properties": {
          "text": { "type": "string" },
          "tag": { "type": "string" }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "key_prefix": { "type": "string" },
          "last_used_at": { "type": "string", "format": "date-time", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "per_page": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
    "/interviews": {
      "get": {
        "summary": "List interviews",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "per_page", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "List of interviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": { "type": "array", "items": { "$ref": "#/components/schemas/Interview" } },
                    "pagination": { "$ref": "#/components/schemas/Pagination" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an interview",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["customer_company", "product_name"],
                "properties": {
                  "customer_company": { "type": "string" },
                  "product_name": { "type": "string" },
                  "customer_email": { "type": "string" },
                  "interview_tone": { "type": "string", "enum": ["formal", "conversational", "technical"] },
                  "interview_focus": { "type": "string", "enum": ["balanced", "roi", "technical", "storytelling"] },
                  "target_audience": { "type": "string", "enum": ["general", "c_suite", "technical_buyer", "end_user", "board"] },
                  "question_limit": { "type": "integer", "minimum": 5, "maximum": 30 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Interview created",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Interview" } } }
              }
            }
          }
        }
      }
    },
    "/interviews/{id}": {
      "get": {
        "summary": "Get interview details",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Interview details", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Interview" } } } } } } }
      },
      "patch": {
        "summary": "Update interview",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "customer_company": { "type": "string" }, "product_name": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Updated interview" } }
      },
      "delete": {
        "summary": "Delete interview",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Deleted", "content": { "application/json": { "schema": { "type": "object", "properties": { "success": { "type": "boolean" } } } } } } }
      }
    },
...
    "/interviews/{id}/draft": {
      "get": {
        "summary": "Get draft content",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Draft content" } }
      },
      "put": {
        "summary": "Update draft content",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["content"], "properties": { "content": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Draft saved" } }
      }
    },
    "/interviews/{id}/formats": {
      "get": {
        "summary": "List generated formats",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Generated formats" } }
      },
      "post": {
        "summary": "Generate a format",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["format"], "properties": { "format": { "type": "string", "enum": ["one_pager", "linkedin", "twitter", "sales_slide", "quote_cards", "email_blurb", "all"] } } } } } },
        "responses": { "200": { "description": "Generated format content" } }
      }
    },
...
    "/interviews/{id}/export": {
      "get": {
        "summary": "Export draft in various formats",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "format", "in": "query", "schema": { "type": "string", "enum": ["md", "docx", "pdf", "html", "txt"], "default": "md" } }
        ],
        "responses": { "200": { "description": "Exported file (binary or text)" } }
      }
    },
...
    "/api-keys": {
      "get": {
        "summary": "List API keys",
        "responses": { "200": { "description": "API keys", "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKey" } } } } } } } }
      },
      "post": {
        "summary": "Create API key (raw key returned once)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string" } } } } } },
        "responses": { "201": { "description": "API key created with raw key" } }
      }
    },
...
  }
}